pyproject.toml
[project]
name = "desk_research"
version = "0.1.0"
description = "Desk Research - Sistema Integrado de Pesquisa AMBEV"
authors = [{ name = "AMBEV Tech", email = "tech@ambev.com.br" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=1.7.0",
    "fpdf2>=2.7.0",
    "python-dotenv>=1.0.0",
    "arxiv>=2.0.0",
    "beautifulsoup4>=4.12.0",
    "feedparser>=6.0.0",
    "pdfplumber>=0.10.0",
    "PyPDF2>=3.0.0",
    "trafilatura>=1.6.0",
    "markdown2>=2.4.0",
    "weasyprint>=60.0.0",
]

[project.scripts]
kickoff = "desk_research.main:kickoff"
run_crew = "desk_research.main:kickoff"
plot = "desk_research.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
